BitfieldField:
- ast: nodes: changing beginBit_ and endBit_ to signed integers so we can more easily detect out of range fields (0 is a valid first field).
- parser: states: code changes to check the bit field value is within the range of the declared underlying type.

+++ swizzle/ast/nodes/BitfieldField.hpp
@@ -2,7 +2,7 @@
 #include <swizzle/ast/Node.hpp>
 #include <swizzle/lexer/TokenInfo.hpp>
 
-#include <cstddef>
+#include <cstdint>
 
 namespace swizzle { namespace ast {
     class VisitorInterface;
@@ -36,7 +36,7 @@ namespace swizzle { namespace ast { namespace nodes {
 
         lexer::TokenInfo valueInfo_;
 
-        std::size_t beginBit_;
-        std::size_t endBit_;
+        std::intmax_t beginBit_;
+        std::intmax_t endBit_;
     };
 }}}
